Cleanup code + bug fix.

+++ Class/Helper/DownloadHelper.cs
@@ -1,6 +1,4 @@
-﻿using ProjBobcat.Class.Model;
-using ProjBobcat.Event;
-using System;
+﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.ComponentModel;
@@ -9,6 +7,8 @@ using System.Linq;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using ProjBobcat.Class.Model;
+using ProjBobcat.Event;
 
 namespace ProjBobcat.Class.Helper
 {
@@ -119,6 +119,64 @@ namespace ProjBobcat.Class.Helper
 
         #endregion
 
+        #region 下载一个列表中的文件（自动确定是否使用分片下载）
+
+        /// <summary>
+        ///     下载文件方法（自动确定是否使用分片下载）
+        /// </summary>
+        /// <param name="fileEnumerable">文件列表</param>
+        /// <param name="downloadThread">下载线程</param>
+        /// <param name="tokenSource"></param>
+        public static async Task AdvancedDownloadListFile(IEnumerable<DownloadFile> fileEnumerable, int downloadThread,
+            CancellationTokenSource tokenSource)
+        {
+            var downloadFiles = fileEnumerable.ToList();
+            var token = tokenSource?.Token ?? CancellationToken.None;
+            var processorCount = ProcessorHelper.GetPhysicalProcessorCount();
+
+            if (downloadThread <= 0)
+                downloadThread = processorCount;
+
+            using var bc = new BlockingCollection<DownloadFile>(downloadThread * 4);
+            using var downloadQueueTask = Task.Run(() =>
+            {
+                foreach (var df in downloadFiles) bc.Add(df, token);
+
+                bc.CompleteAdding();
+            }, token);
+
+            using var downloadTask = Task.Run(() =>
+            {
+                void DownloadAction()
+                {
+                    foreach (var df in bc.GetConsumingEnumerable())
+                    {
+                        var di = new DirectoryInfo(
+                            df.DownloadPath.Substring(0, df.DownloadPath.LastIndexOf('\\')));
+                        if (!di.Exists) di.Create();
+
+                        // DownloadData(df);
+                        if (df.FileSize >= 1048576 || df.FileSize == 0 || df.FileSize == default)
+                            MultiPartDownload(df);
+                        else
+                            DownloadData(df);
+                    }
+                }
+
+                var threads = new List<Thread>();
+
+                for (var i = 0; i < downloadThread * 2; i++) threads.Add(new Thread(DownloadAction));
+
+                foreach (var t in threads) t.Start();
+
+                foreach (var t in threads) t.Join();
+            }, token);
+
+            await Task.WhenAll(downloadQueueTask, downloadTask).ConfigureAwait(false);
+        }
+
+        #endregion
+
         #region 分片下载
 
         public static void MultiPartDownload(DownloadFile downloadFile, int numberOfParts = 16)
@@ -126,27 +184,20 @@ namespace ProjBobcat.Class.Helper
             if (downloadFile == null) return;
 
             //Handle number of parallel downloads  
-            if (numberOfParts <= 0)
-            {
-                numberOfParts = Environment.ProcessorCount;
-            }
+            if (numberOfParts <= 0) numberOfParts = Environment.ProcessorCount;
 
             try
             {
                 #region Get file size
 
-                var webRequest = (HttpWebRequest)WebRequest.Create(new Uri(downloadFile.DownloadUri));
+                var webRequest = (HttpWebRequest) WebRequest.Create(new Uri(downloadFile.DownloadUri));
                 webRequest.Method = "HEAD";
                 webRequest.UserAgent = Ua;
-                long responseLength;
-                bool parallelDownloadSupported;
 
-                using (var webResponse = webRequest.GetResponse())
-                {
-                    parallelDownloadSupported = webResponse.Headers.Get("Accept-Ranges")?.Contains("bytes") ?? false;
-                    responseLength = long.TryParse(webResponse.Headers.Get("Content-Length"), out var l) ? l : 0;
-                    parallelDownloadSupported = (parallelDownloadSupported && responseLength != 0);
-                }
+                using var webResponse = webRequest.GetResponse();
+                var parallelDownloadSupported = webResponse.Headers.Get("Accept-Ranges")?.Contains("bytes") ?? false;
+                var responseLength = long.TryParse(webResponse.Headers.Get("Content-Length"), out var l) ? l : 0;
+                parallelDownloadSupported = parallelDownloadSupported && responseLength != 0;
 
                 if (!parallelDownloadSupported)
                 {
@@ -156,17 +207,14 @@ namespace ProjBobcat.Class.Helper
 
                 #endregion
 
-                if (File.Exists(downloadFile.DownloadPath))
-                {
-                    File.Delete(downloadFile.DownloadPath);
-                }
+                if (File.Exists(downloadFile.DownloadPath)) File.Delete(downloadFile.DownloadPath);
 
                 var tempFilesDictionary = new ConcurrentDictionary<int, byte[]>();
 
                 #region Calculate ranges
 
                 var readRanges = new List<DownloadRange>();
-                var partSize = (long)Math.Ceiling((double)responseLength / numberOfParts);
+                var partSize = (long) Math.Ceiling((double) responseLength / numberOfParts);
 
                 for (var i = 0; i < numberOfParts; i++)
                 {
@@ -189,7 +237,7 @@ namespace ProjBobcat.Class.Helper
 
                 Task.WhenAll(readRanges.Select(range => Task.Run(() =>
                 {
-                    using var client = new WebClient { DownloadRange = range };
+                    using var client = new WebClient {DownloadRange = range};
                     client.Headers.Add("user-agent", Ua);
 
                     var data = client.DownloadData(new Uri(downloadFile.DownloadUri));
@@ -197,7 +245,7 @@ namespace ProjBobcat.Class.Helper
                     downloadParts++;
 
                     downloadFile.Changed?.Invoke(client,
-                        new DownloadFileChangedEventArgs { ProgressPercentage = (double)downloadParts / numberOfParts });
+                        new DownloadFileChangedEventArgs {ProgressPercentage = (double) downloadParts / numberOfParts});
                 })).ToArray());
 
                 #endregion
@@ -214,9 +262,7 @@ namespace ProjBobcat.Class.Helper
                 }
 
                 foreach (var downloadedBytes in tempFilesDictionary.OrderBy(b => b.Key))
-                {
                     bytes.AddRange(downloadedBytes.Value);
-                }
 
                 if (bytes.Count != responseLength)
                 {
@@ -245,10 +291,7 @@ namespace ProjBobcat.Class.Helper
             if (downloadFile == null) return;
 
             //Handle number of parallel downloads  
-            if (numberOfParts <= 0)
-            {
-                numberOfParts = Environment.ProcessorCount;
-            }
+            if (numberOfParts <= 0) numberOfParts = Environment.ProcessorCount;
 
             try
             {
@@ -264,7 +307,7 @@ namespace ProjBobcat.Class.Helper
                 {
                     parallelDownloadSupported = webResponse.Headers.Get("Accept-Ranges")?.Contains("bytes") ?? false;
                     responseLength = long.TryParse(webResponse.Headers.Get("Content-Length"), out var l) ? l : 0;
-                    parallelDownloadSupported = (parallelDownloadSupported && responseLength != 0);
+                    parallelDownloadSupported = parallelDownloadSupported && responseLength != 0;
                 }
 
                 if (!parallelDownloadSupported)
@@ -275,17 +318,14 @@ namespace ProjBobcat.Class.Helper
 
                 #endregion
 
-                if (File.Exists(downloadFile.DownloadPath))
-                {
-                    File.Delete(downloadFile.DownloadPath);
-                }
+                if (File.Exists(downloadFile.DownloadPath)) File.Delete(downloadFile.DownloadPath);
 
                 var tempFilesDictionary = new ConcurrentDictionary<int, byte[]>();
 
                 #region Calculate ranges
 
                 var readRanges = new List<DownloadRange>();
-                var partSize = (long)Math.Ceiling((double)responseLength / numberOfParts);
+                var partSize = (long) Math.Ceiling((double) responseLength / numberOfParts);
 
                 for (var i = 0; i < numberOfParts; i++)
                 {
@@ -313,7 +353,7 @@ namespace ProjBobcat.Class.Helper
                     var data = await client.DownloadDataTaskAsync(new Uri(downloadFile.DownloadUri))
                         .ConfigureAwait(false);
 
-                    if(!tempFilesDictionary.TryAdd(range.Index, data)) return;
+                    if (!tempFilesDictionary.TryAdd(range.Index, data)) return;
                     downloadParts++;
 
                     downloadFile.Changed?.Invoke(client,
@@ -334,9 +374,7 @@ namespace ProjBobcat.Class.Helper
                 }
 
                 foreach (var downloadedBytes in tempFilesDictionary.OrderBy(b => b.Key))
-                {
                     bytes.AddRange(downloadedBytes.Value);
-                }
 
                 if (bytes.Count != responseLength)
                 {
@@ -358,80 +396,6 @@ namespace ProjBobcat.Class.Helper
                 downloadFile.Completed?.Invoke(null,
                     new DownloadFileCompletedEventArgs(false, ex, downloadFile));
             }
-        }
-
-        #endregion
-
-        #region 下载一个列表中的文件（自动确定是否使用分片下载）
-
-        /// <summary>
-        ///     下载文件方法（自动确定是否使用分片下载）
-        /// </summary>
-        /// <param name="fileEnumerable">文件列表</param>
-        /// <param name="downloadThread">下载线程</param>
-        /// <param name="tokenSource"></param>
-        public static async Task AdvancedDownloadListFile(IEnumerable<DownloadFile> fileEnumerable, int downloadThread,
-        CancellationTokenSource tokenSource)
-        {
-            var downloadFiles = fileEnumerable.ToList();
-            var token = tokenSource?.Token ?? CancellationToken.None;
-            var processorCount = ProcessorHelper.GetPhysicalProcessorCount();
-
-            if (downloadThread <= 0)
-                downloadThread = processorCount;
-
-            using var bc = new BlockingCollection<DownloadFile>(downloadThread * 4);
-            using var downloadQueueTask = Task.Run(() =>
-            {
-                foreach (var df in downloadFiles)
-                {
-                    bc.Add(df, token);
-                }
-
-                bc.CompleteAdding();
-            }, token);
-
-            using var downloadTask = Task.Run(() =>
-            {
-                void DownloadAction()
-                {
-                    foreach (var df in bc.GetConsumingEnumerable())
-                    {
-                        var di = new DirectoryInfo(
-                            df.DownloadPath.Substring(0, df.DownloadPath.LastIndexOf('\\')));
-                        if (!di.Exists) di.Create();
-
-                        // DownloadData(df);
-                        if (df.FileSize >= 1048576 || df.FileSize == 0 || df.FileSize == default)
-                        {
-                            MultiPartDownload(df);
-                        }
-                        else
-                        {
-                            DownloadData(df);
-                        }
-                    }
-                }
-
-                var threads = new List<Thread>();
-
-                for (var i = 0; i < downloadThread * 2; i++)
-                {
-                    threads.Add(new Thread(DownloadAction));
-                }
-
-                foreach (var t in threads)
-                {
-                    t.Start();
-                }
-
-                foreach (var t in threads)
-                {
-                    t.Join();
-                }
-            }, token);
-
-            await Task.WhenAll(downloadQueueTask, downloadTask).ConfigureAwait(false);
         }
 
         #endregion
